Fix issues with frequency
- For annual frequency,
pandas understand  "Y" and "A", but only "A" can have an anchor.
- When the frequency is anchored, for example "W-WED",
we have to make sure it is a anchor we can reuse for bootstrap,
before picking it.
- Also added to parse_offset the "S" stamp in its own group
- When running bootstrap_year on a anchored frequency,
the bloc and the source may have different lengths on the edges of
the time series.
  - when bloc length is > to source length we can only ignore the bloc
  - when bloc length is < to source length, we can copy a part source
Plus, there were cases were a duplication could occurred for the first
days of the year.