ENH: Improve csdi/wdsi descriptions
Previously, their definitions were possibly incorrect in the case
where the user would compute percentiles with custom parameters
using `percentile_doy`.

This concerns, the base percentile value, the length of rolling
window of days and the period on which percentiles are computed.

A limit of the current approach is that the percentiles **must**
be computed using `percentile_doy` because some metadata are
expected in `DataArray.attrs`.