Add JOSS Article, Review badge and Citation information (#250)
### What kind of change does this PR introduce?

This PR finalizes the submission and publishing steps for the [Journal
of Open Source
Software](https://joss.readthedocs.io/en/latest/index.html). The paper
is expected to have a length of 250 - 1000 words and demands that the
software is *feature-complete*. As such, the submission process should
not be started until we have at the very least added all necessary
indicators from our backlog and/or stabilized our API (i.e.: v1.0-alpha
or release-candidate).

Updates (May 2023): 
With xclim v0.40.0, the software was deemed ready for submission. The
review process for JOSS was completed via PyOpenSci
(https://github.com/pyOpenSci/software-submission/issues/73), and the
final JOSS review was performed in
https://github.com/openjournals/joss-reviews/issues/5415. The software
(v0.43.0) and paper were published on 18 May 2023
(https://doi.org/10.21105/joss.05415).

### Does this PR introduce a breaking change?

No.

---------

Co-authored-by: Philippe Roy <borghor@yahoo.ca>
Co-authored-by: David Huard <huard.david@ouranos.ca>
Co-authored-by: Abel Aoun <aoun@cerfacs.fr>
Co-authored-by: Pascal Bourgault <bourgault.pascal@ouranos.ca>
Co-authored-by: tlogan2000 <logan.travis@ouranos.ca>